Move function definitions from session.h to session.c to decouple SDK deps
Root cause: session.h defined alloc_shared_mem_buf and free_shared_mem_buf
with full function bodies that reference FastRPC symbols (rpcmem_alloc,
fastrpc_mmap, etc). Including this header from float_tensor.cpp caused
these SDK symbols to appear in the object file, requiring libcdsprpc.so
at link time - which is only available on the target device.

Fix:
- session.h now contains only declarations (no SDK header includes needed)
- Function definitions moved to session.c (compiled via cmake for target)
- Added extern "C" guards for safe C++ inclusion
- Added guarded remote_handle64 typedef to avoid SDK dependency
- meson.build reverted to original custom_target approach (cmake handles
  SDK dependencies for target deployment)
- Removed htp_dep from nntrainer_base_deps (not needed at build time)
