+++ nntrainer/tensor/htp_backend/host/session.c
@@ -1,4 +1,6 @@
 #include <AEEStdErr.h>
+#include <remote.h>
+#include <rpcmem.h>
 #include <stdio.h>
 #include <stdlib.h>
 
@@ -93,4 +95,33 @@ void init_htp_backend() {
 
 int create_htp_message_channel(int fd, unsigned int max_msg_size) {
   return htp_ops_create_channel(session_handle, fd, max_msg_size);
+}
+
+int alloc_shared_mem_buf(void **p_buf, int *p_fd, size_t size) {
+  void *buf = rpcmem_alloc(RPCMEM_HEAP_ID_SYSTEM, RPCMEM_FLAG_UNCACHED, size);
+  if (!buf) {
+    fprintf(stderr, "alloc_shared_mem_buf: rpcmem_alloc failed\n");
+    return -1;
+  }
+
+  int fd = rpcmem_to_fd(buf);
+  if (fd < 0) {
+    fprintf(stderr, "alloc_shared_mem_buf: rpcmem_to_fd failed\n");
+    return -1;
+  }
+
+  int err = fastrpc_mmap(CDSP_DOMAIN_ID, fd, buf, 0, size, FASTRPC_MAP_FD);
+  if (err) {
+    fprintf(stderr, "alloc_shared_mem_buf: fastrpc_mmap failed, err: %d\n", err);
+    return -1;
+  }
+
+  *p_buf = buf;
+  *p_fd  = fd;
+  return 0;
+}
+
+void free_shared_mem_buf(void *buf, int fd, size_t size) {
+  fastrpc_munmap(CDSP_DOMAIN_ID, fd, buf, size);
+  rpcmem_free(buf);
 }
