+++ nntrainer/tensor/htp_backend/include/host/session.h
@@ -1,50 +1,28 @@
 #pragma once
 
-#include <remote.h>
-#include <rpcmem.h>
 #include <stdio.h>
 #include <stdint.h>
-#include <stdio.h>
 #include <stdlib.h>
-#include <time.h>
+#include <stddef.h>
 #include <unistd.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#ifndef REMOTE_HANDLE64_TYPEDEF
+#define REMOTE_HANDLE64_TYPEDEF
+typedef uint64_t remote_handle64;
+#endif
 
 int open_dsp_session(int domain_id, int unsigned_pd_enabled);
 void close_dsp_session();
-
 remote_handle64 get_global_handle();
-
 void init_htp_backend();
 int create_htp_message_channel(int fd, unsigned int max_msg_size);
+int alloc_shared_mem_buf(void **p_buf, int *p_fd, size_t size);
+void free_shared_mem_buf(void *buf, int fd, size_t size);
 
-
-// assert p_buf, p_fd and size are always valid
-int alloc_shared_mem_buf(void **p_buf, int *p_fd, size_t size) {
-  void *buf = rpcmem_alloc(RPCMEM_HEAP_ID_SYSTEM, RPCMEM_FLAG_UNCACHED, size);
-  if (!buf) {
-    fprintf(stderr, "alloc_shared_mem_buf: rpcmem_alloc failed\n");
-    return -1;
-  }
-
-  int fd = rpcmem_to_fd(buf);
-  if (fd < 0) {
-    fprintf(stderr, "alloc_shared_mem_buf: rpcmem_to_fd failed\n");
-    return -1;
-  }
-
-  // map buffer to the DSP
-  int err = fastrpc_mmap(CDSP_DOMAIN_ID, fd, buf, 0, size, FASTRPC_MAP_FD);
-  if (err) {
-    fprintf(stderr, "alloc_shared_mem_buf: fastrpc_mmap failed, err: %d\n", err);
-    return -1;
-  }
-
-  *p_buf = buf;
-  *p_fd  = fd;
-  return 0;
+#ifdef __cplusplus
 }
-
-void free_shared_mem_buf(void *buf, int fd, size_t size) {
-  fastrpc_munmap(CDSP_DOMAIN_ID, fd, buf, size);
-  rpcmem_free(buf);
-}
+#endif
